requirements.txt
amqp==5.1.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
kombu==5.2.4
lxml==4.8.0
multitasking==0.0.10
numpy==1.22.3
packaging==21.3
pandas==1.4.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
yfinance==0.1.70
